Some test fixes
