+++ instrumentation-api-annotation-support/src/test/groovy/io/opentelemetry/instrumentation/api/annotation/support/MethodSpanAttributesExtractorTest.groovy
@@ -6,7 +6,7 @@
 package io.opentelemetry.instrumentation.api.annotation.support
 
 import io.opentelemetry.api.common.AttributesBuilder
-import io.opentelemetry.instrumentation.api.caching.Cache
+import io.opentelemetry.instrumentation.api.cache.Cache
 import spock.lang.Specification
 
 import java.lang.reflect.Method
